requirements.txt
tornado==6.4
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.22
langchain-openai==0.0.5
langchain-pinecone==0.0.2
python-dotenv==1.0.0
